Remove settings.conf from the repo
settings.conf shouldn't be in the repo, because it's ~always going to be edited be the user.

We don't need it there as an example, because we _have_ and example in
the form of settings.conf.example, and running kandji2snipe w/o a
settings.conf file gives a helpful response.

On top of all that,  the settings.conf file that was there was out of date -- missing many
   of the required fields and all of the default values displayed in
   settings.conf.example

I've also added settings.conf to .gitignore, so that the repo remain
clean when people inevitably create a settings.conf file